Refactor MockMvcExtensions.kt
